#1_home_news_fragment: Auto stash before merge of "feat/#1_home_news_fragment" and "origin/feat/#1_home_news_fragment"

+++ app/src/main/java/com/umc/ttoklip/presentation/hometown/ReadCommunicationActivity.kt
@@ -223,7 +223,7 @@ class ReadCommunicationActivity :
 
     companion object {
         const val COMMUNICATION = "postId"
-        fun newIntent(context: Context, id: Int) =
+        fun newIntent(context: Context, id: Long) =
             Intent(context, ReadCommunicationActivity::class.java).apply {
                 putExtra(COMMUNICATION, id)
             }
